BIG UPDATE: ADDED A BOT
In this update I made a basic bot that is very easy to beat, I may add another one, but I dont know yet. Basically it just has a list of locations in an order that it goes, if you want to test it just go on the top row to win. HAVE FUN!